+++ Patches/InitializeGamePatch.cs
@@ -1,6 +1,5 @@
 using HarmonyLib;
 using IntroTweaks.Utils;
-using UnityEngine.SceneManagement;
 
 namespace IntroTweaks.Patches;
 
@@ -8,15 +7,16 @@ namespace IntroTweaks.Patches;
 internal class InitializeGamePatch {
     [HarmonyPrefix]
     [HarmonyPatch("Start")]
-    static void DisableBootAnimation(ref bool __runOriginal) {
+    static void DisableBootAnimation(InitializeGame __instance) {
         int startupDisplayIndex = Plugin.Config.GAME_STARTUP_DISPLAY.Value;
         if (startupDisplayIndex >= 0) {
             DisplayUtil.Move(startupDisplayIndex);
         }
 
         if (Plugin.Config.SKIP_BOOT_ANIMATION.Value) {
-            SceneManager.LoadScene("MainMenu");
-            __runOriginal = false;
+            __instance.runBootUpScreen = false;
+            __instance.bootUpAudio = null;
+            __instance.bootUpAnimation = null;
         }
     }
 }